+++ caoslib/initialization/__init__.py
@@ -3,8 +3,7 @@ from clint.textui import puts, colored
 
 options = ['init']
 
-def handler(mode):
+def handler(mode, args):
     init()
     puts(colored.green("Init successful"))
-
     exit(0)
